Move python file from python/src to main reputation folder
