+++ reputation/__init__.py
@@ -1,5 +1,5 @@
 name = "reputation"
 
-from reputation.python.src.aigents_reputation_api import AigentsAPIReputationService
+from reputation.aigents_reputation_api import AigentsAPIReputationService
