+++ reputation/test_reputation_simulation.py
@@ -42,7 +42,7 @@ class TestReputationSimulationBase(object):
 		bad_agent = {"range": [9,10], "values": [1,10], "transactions": 100, "suppliers": 1, "consumers": 1}
 		reputation_simulate(good_agent,bad_agent, datetime.date(2018, 1, 1), 10, True, None, False)
 		#Step 2 - process simulated with reputaion engine in batch mode, grab results and check them
-		cmd = 'python reputation_simulate.py ../../bin testsim ./ transactions10_r_100_0.1.tsv users10.tsv 2018-01-01 2018-01-10 logratings=False weighting=True fullnorm=True default=0.5'
+		cmd = 'python reputation_simulate.py ./bin testsim ./ transactions10_r_100_0.1.tsv users10.tsv 2018-01-01 2018-01-10 logratings=False weighting=True fullnorm=True default=0.5'
 		r = subprocess.check_output(cmd,shell=True)
 		lines = r.decode().splitlines()
 		#print(lines)
@@ -56,7 +56,7 @@ class TestReputationSimulationBase(object):
 		bad_agent = {"range": [9,10], "values": [1,10], "transactions": 100, "suppliers": 1, "consumers": 1}
 		reputation_simulate(good_agent,bad_agent, datetime.date(2018, 1, 1), 10, False, None, False)
 		#Step 2 - process simulated with reputaion engine in batch mode, grab results and check them
-		cmd = 'python reputation_simulate.py ../../bin testsim ./ transactions10_p_100_0.1.tsv users10.tsv 2018-01-01 2018-01-10 logratings=False weighting=True fullnorm=True default=0.5'
+		cmd = 'python reputation_simulate.py ./bin testsim ./ transactions10_p_100_0.1.tsv users10.tsv 2018-01-01 2018-01-10 logratings=False weighting=True fullnorm=True default=0.5'
 		r = subprocess.check_output(cmd,shell=True)
 		#os.system(cmd)
 		lines = r.decode().splitlines()
@@ -91,7 +91,7 @@ class TestReputationSimulationAigents(TestReputationSimulationBase,unittest.Test
 
 	@classmethod
 	def setUpClass(cls):
-		cmd = 'java -cp ../../bin/Aigents.jar:../../bin/* net.webstructor.agent.Farm store path \'./al_test.txt\', http port 1180, cookie domain localtest.com, console off'
+		cmd = 'java -cp ./bin/Aigents.jar:./bin/* net.webstructor.agent.Farm store path \'./al_test.txt\', http port 1180, cookie domain localtest.com, console off'
 		cls.server_process = subprocess.Popen(cmd.split())
 		#self.server_process = subprocess.Popen(['sh','aigents_server_start.sh'])
 		time.sleep(10)
