+++ scripts/test
@@ -3,7 +3,7 @@ set -ex
 
 PARENT_PATH=$(dirname $(cd $(dirname $0); pwd -P))
 
-pushd $PARENT_PATH/reputation/python/src
+pushd $PARENT_PATH/reputation
 
 python3 -m unittest
